Do not update visionOffsetLog, when there is no entries.
When the varianceHistory is set to 0, we want to keep the vision enabled
at all time. In such cases, there is no need to log anything and no log
entries are allocated, and thus we short-circuit the code in charge of
updating the log.

This caused some problem, as we could not remove the first element of a
non allocated container.